# infrastructure/swagger/openapi.yml
openapi: 3.0.3
info:
  title: Smart Banking System API
  description: |
    Complete REST API for the TT-BANK Smart Banking System.

    **Architecture:** Microservices — all requests enter through the API Gateway on port 8080.
    The gateway validates the JWT Bearer token before routing to the appropriate backend service.

    **Authentication flow:**
    1. `POST /api/v1/auth/register` — create account
    2. `POST /api/v1/auth/verify-email` — verify OTP sent by email
    3. `POST /api/v1/auth/login` — receive `accessToken` and `refreshToken`
    4. Include `Authorization: Bearer <accessToken>` on all subsequent requests

    **Currency:** Primary currency is **XAF** (Central African CFA franc). Also supports USD, EUR, GBP.

    **Base URL (local dev):** `http://localhost:8080`
    **Base URL (production):** `https://yourdomain.com`
  version: "1.0.0"
  contact:
    name: Smart Banking Dev Team
servers:
  - url: http://localhost:8080
    description: Local development (API Gateway)
  - url: https://yourdomain.com
    description: Production (replace with actual domain)

# =============================================================
# Security scheme — JWT Bearer token (HS256)
# =============================================================
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        JWT access token issued by POST /api/v1/auth/login.
        Include as: Authorization: Bearer <token>

  # =============================================================
  # Reusable schemas — built from actual Java DTOs
  # =============================================================
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
          example: "Operation completed successfully."
...
    WalletResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        userId:
          type: integer
          format: int64
          example: 1
        walletNumber:
          type: string
          example: "WLT-20240601-00001"
        ownerName:
          type: string
          example: "Jean-Pierre Nkomo"
        email:
          type: string
          example: "jeanpierre@example.com"
        phoneNumber:
          type: string
          example: "+237612345678"
        balance:
          type: number
          format: decimal
          example: 125000.00
        currency:
          $ref: '#/components/schemas/Currency'
        status:
          $ref: '#/components/schemas/WalletStatus'
        createdAt:
          type: string
          format: date-time
          example: "2024-06-01T10:00:00"
    DepositRequest:
      type: object
      required: [amount, description]
      properties:
        amount:
          type: number
          format: decimal
          minimum: 1.00
          example: 50000.00
        description:
          type: string
          example: "Monthly salary deposit"
    WithdrawRequest:
      type: object
      required: [amount, description]
      properties:
        amount:
          type: number
          format: decimal
          minimum: 1.00
          example: 10000.00
        description:
          type: string
          example: "ATM withdrawal"
    Currency:
      type: string
      enum: [XAF, USD, EUR, GBP]
    WalletStatus:
      type: string
      enum: [ACTIVE, SUSPENDED, CLOSED]

    # ── Transaction ───────────────────────────────────────────
...
    TransactionResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 42
        referenceCode:
          type: string
          example: "TXN-20240601-ABC123"
        type:
          $ref: '#/components/schemas/TransactionType'
        status:
          $ref: '#/components/schemas/TransactionStatus'
        senderWalletNumber:
          type: string
          example: "WLT-20240601-00001"
        senderEmail:
          type: string
          example: "sender@example.com"
        receiverWalletNumber:
          type: string
          example: "WLT-20240601-00002"
        receiverEmail:
          type: string
          example: "recipient@example.com"
        amount:
          type: number
          format: decimal
          example: 5000.00
        currency:
          type: string
          example: "XAF"
        description:
          type: string
          example: "Payment for services"
        createdAt:
          type: string
          format: date-time
          example: "2024-06-01T14:30:00"
        completedAt:
          type: string
          format: date-time
          nullable: true
          example: "2024-06-01T14:30:01"
    TransactionType:
      type: string
      enum: [TRANSFER, DEPOSIT, WITHDRAWAL, MERCHANT_PAYMENT]
    TransactionStatus:
      type: string
      enum: [PENDING, COMPLETED, FAILED, REVERSED]

    # ── Merchant ──────────────────────────────────────────────
...
    ErrorResponse:
      type: object
      properties:
        status:
          type: integer
          example: 400
        error:
          type: string
          example: "Bad Request"
        message:
          type: string
          example: "Validation failed: email must be a valid email address"
        timestamp:
          type: string
          format: date-time
          example: "2024-06-01T14:30:00"
        path:
          type: string
          example: "/api/v1/auth/register"
  responses:
    Unauthorized:
      description: Missing or invalid JWT Bearer token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Valid token but insufficient role (e.g. non-ADMIN accessing /api/v1/admin/**)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
security:
  - BearerAuth: []

# =============================================================
# API Paths
# =============================================================
paths:
...
  /api/v1/wallet:
    post:
      tags: [Wallet]
      summary: Create a wallet for the authenticated user
      description: Creates a new XAF wallet linked to the authenticated user. Each user can only have one wallet.
      responses:
        '201':
          description: Wallet created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WalletResponse'
        '409':
          description: User already has a wallet
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/wallet/create:
    post:
      tags: [Wallet]
      summary: Create wallet (alternate endpoint)
      description: Identical to POST /api/v1/wallet — provided for REST consistency.
      responses:
        '201':
          description: Wallet created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WalletResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/wallet/me:
    get:
      tags: [Wallet]
      summary: Get the authenticated user's wallet
      responses:
        '200':
          description: Wallet details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WalletResponse'
        '404':
          description: User does not have a wallet
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/wallet/number/{walletNumber}:
    get:
      tags: [Wallet]
      summary: Look up wallet by wallet number
      parameters:
        - name: walletNumber
          in: path
          required: true
          schema:
            type: string
            example: "WLT-20240601-00001"
      responses:
        '200':
          description: Wallet found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WalletResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/wallet/phone/{phoneNumber}:
    get:
      tags: [Wallet]
      summary: Look up wallet by phone number
      parameters:
        - name: phoneNumber
          in: path
          required: true
          schema:
            type: string
            example: "+237612345678"
      responses:
        '200':
          description: Wallet found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WalletResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/wallet/deposit:
    post:
      tags: [Wallet]
      summary: Deposit funds into the authenticated user's wallet
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DepositRequest'
      responses:
        '200':
          description: Deposit successful — returns updated wallet
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WalletResponse'
        '400':
          description: Amount exceeds maximum deposit limit (10,000,000 XAF)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/wallet/withdraw:
    post:
      tags: [Wallet]
      summary: Withdraw funds from the authenticated user's wallet
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WithdrawRequest'
      responses:
        '200':
          description: Withdrawal successful — returns updated wallet
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WalletResponse'
        '400':
          description: Insufficient funds, amount exceeds limit, or daily limit reached
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/wallet/transactions:
    get:
      tags: [Wallet]
      summary: Get wallet transaction history
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Paginated list of wallet transactions
          content:
            application/json:
              schema:
                type: object
                properties:
                  content:
                    type: array
                    items:
                      $ref: '#/components/schemas/TransactionResponse'
                  totalElements:
                    type: integer
                    example: 142
                  totalPages:
                    type: integer
                    example: 8
                  number:
                    type: integer
                    example: 0
                  size:
                    type: integer
                    example: 20
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/wallet/admin/{walletNumber}/suspend:
    post:
      tags: [Wallet, Admin]
      summary: Suspend a wallet (Admin only)
      parameters:
        - name: walletNumber
          in: path
          required: true
          schema:
            type: string
            example: "WLT-20240601-00001"
      responses:
        '200':
          description: Wallet suspended
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/wallet/admin/{walletNumber}/reactivate:
    post:
      tags: [Wallet, Admin]
      summary: Reactivate a suspended wallet (Admin only)
      parameters:
        - name: walletNumber
          in: path
          required: true
          schema:
            type: string
            example: "WLT-20240601-00001"
      responses:
        '200':
          description: Wallet reactivated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

  # ============================================================
  # TRANSACTION SERVICE — /api/v1/transactions/**
  # ============================================================
...